Advertencia Box para eliminar cliente

+++ Veterinaria-Login/ControlClientes/EliminarCliente.cs
@@ -23,52 +23,59 @@ namespace Veterinaria_Login
                     throw new Exception("Debe ingresar un DNI válido.");
                 }
 
-                using (SqlConnection conn = new SqlConnection("Data Source=DESKTOP-SLVGAGD\\SQLEXPRESS;Initial Catalog=Sistema2;Integrated Security=True"))
+                DialogResult result = MessageBox.Show("¿Estás seguro de que deseas eliminar al cliente y todos los registros relacionados?", "Advertencia", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+                if (result == DialogResult.Yes)
                 {
-                    conn.Open();
-
-                    string deleteMascotasQuery = "DELETE FROM Mascotas WHERE DniCliente = @dni";
-                    using (SqlCommand deleteMascotasCommand = new SqlCommand(deleteMascotasQuery, conn))
+                    using (SqlConnection conn = new SqlConnection("Data Source=DESKTOP-SLVGAGD\\SQLEXPRESS;Initial Catalog=Sistema2;Integrated Security=True"))
                     {
-                        deleteMascotasCommand.Parameters.AddWithValue("@dni", dni);
-                        deleteMascotasCommand.ExecuteNonQuery();
-                    }
+                        conn.Open();
 
-                    string deleteConsultasQuery = "DELETE FROM Consultas WHERE DniCliente = @dni";
-                    using (SqlCommand deleteConsultasCommand = new SqlCommand(deleteConsultasQuery, conn))
-                    {
-                        deleteConsultasCommand.Parameters.AddWithValue("@dni", dni);
-                        deleteConsultasCommand.ExecuteNonQuery();
-                    }
-
-
-                    string deleteTratamientosQuery = "DELETE FROM Tratamientos WHERE DniDueñoMascota = @dni";
-                    using (SqlCommand deleteTratamientosCommand = new SqlCommand(deleteTratamientosQuery, conn))
-                    {
-                        deleteTratamientosCommand.Parameters.AddWithValue("@dni", dni);
-                        deleteTratamientosCommand.ExecuteNonQuery();
-                    }
-
-
-                    string deleteClientesQuery = "DELETE FROM Clientes WHERE Dni = @dni";
-                    using (SqlCommand deleteClientesCommand = new SqlCommand(deleteClientesQuery, conn))
-                    {
-                        deleteClientesCommand.Parameters.AddWithValue("@dni", dni);
-
-                        int rowsAffected = deleteClientesCommand.ExecuteNonQuery();
-
-                        if (rowsAffected > 0)
+                        string deleteMascotasQuery = "DELETE FROM Mascotas WHERE DniCliente = @dni";
+                        using (SqlCommand deleteMascotasCommand = new SqlCommand(deleteMascotasQuery, conn))
                         {
-                            MessageBox.Show("Persona y registros relacionados eliminados correctamente");
+                            deleteMascotasCommand.Parameters.AddWithValue("@dni", dni);
+                            deleteMascotasCommand.ExecuteNonQuery();
                         }
-                        else
-                        {
-                            MessageBox.Show("No se encontró ninguna persona con el DNI especificado");
-                        }
-                    }
 
-                    conn.Close();
+                        string deleteConsultasQuery = "DELETE FROM Consultas WHERE DniCliente = @dni";
+                        using (SqlCommand deleteConsultasCommand = new SqlCommand(deleteConsultasQuery, conn))
+                        {
+                            deleteConsultasCommand.Parameters.AddWithValue("@dni", dni);
+                            deleteConsultasCommand.ExecuteNonQuery();
+                        }
+
+
+                        string deleteTratamientosQuery = "DELETE FROM Tratamientos WHERE DniDueñoMascota = @dni";
+                        using (SqlCommand deleteTratamientosCommand = new SqlCommand(deleteTratamientosQuery, conn))
+                        {
+                            deleteTratamientosCommand.Parameters.AddWithValue("@dni", dni);
+                            deleteTratamientosCommand.ExecuteNonQuery();
+                        }
+
+
+                        string deleteClientesQuery = "DELETE FROM Clientes WHERE Dni = @dni";
+                        using (SqlCommand deleteClientesCommand = new SqlCommand(deleteClientesQuery, conn))
+                        {
+                            deleteClientesCommand.Parameters.AddWithValue("@dni", dni);
+
+                            int rowsAffected = deleteClientesCommand.ExecuteNonQuery();
+
+                            if (rowsAffected > 0)
+                            {
+                                MessageBox.Show("Cliente y registros relacionados eliminados correctamente.");
+                                DelCliente.Text = "";
+                            }
+                            else
+                            {
+                                MessageBox.Show("No se encontró ningun Cliente con el DNI especificado.");
+                                DelCliente.Text = "";
+                            }
+                        }
+
+                        conn.Close();
+                    }
                 }
+
             }
             catch (Exception ex)
             {
